pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[project]
name = "eta-engine"
version = "0.1.0"
description = "4-bot profit-recycling organism: MNQ/NQ -> Crypto Seed -> ETH/SOL/XRP Perps + Staking"
authors = [{name = "Edward Avila", email = "edward.t.avila@gmail.com"}]
requires-python = ">=3.12"
dependencies = [
    # Core
    "pydantic>=2.0",
    "numpy>=1.26",
    "pandas>=2.1",
    "polars>=0.20",

    # Data feeds
    "ccxt>=4.0",
    "websockets>=12.0",

    # Storage
    "arcticdb>=4.0",
    "pyarrow>=14.0",

    # Broker adapters
    "httpx>=0.27",
    "aiohttp>=3.9",

    # ML / RL
    "torch>=2.2",
    "stable-baselines3>=2.2",
    "scikit-learn>=1.4",
    "lightgbm>=4.2",
    "river>=0.21",

    # Crypto / DeFi
    "web3>=6.15",
    "solana>=0.32",

    # Monitoring
    "prometheus-client>=0.20",
    "python-telegram-bot>=21.0",

    # Infra
    "fastapi>=0.110",
    "uvicorn>=0.27",
    "redis>=5.0",
    "psycopg[binary]>=3.1",

    # Auth / file locking
    "bcrypt>=4.1",
    "portalocker>=2.10",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "ruff>=0.3",
    "mypy>=1.8",
]
litellm = ["litellm>=1.50.0"]
mem0 = ["mem0ai>=0.1.0"]
eval = ["promptfoo>=0.100.0"]
mlflow = ["mlflow>=2.15.0"]
langfuse = ["langfuse>=2.50.0"]
openrouter = []  # uses openai SDK (already available as transitive)

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "W", "F", "I", "N", "UP", "ANN", "B", "SIM", "TCH"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ANN001", "ANN002", "ANN003", "ANN201", "ANN202", "ANN401", "N802", "N803", "TC003"]
# jarvis_dashboard embeds long HTML template strings (E501) and overrides
# BaseHTTPRequestHandler.log_message whose signature uses *args: Any (ANN401).
"scripts/jarvis_dashboard.py" = ["E501", "ANN401"]
"scripts/jarvis_status_server.py" = ["E501"]
# run_eta_live is a dispatch-heavy tie-together runtime: the router, bot
# factories, and OrderRequest surface cross module boundaries where a Protocol
# or Union would require pulling in 6+ venue/bot imports. The Any types are
# intentional extension points. *_ / **__ are placeholder param names for
# signatures that must match a venue-supplied contract whose extras we don't
# use.
"scripts/run_eta_live.py" = ["ANN002", "ANN003", "ANN401"]
"feeds/run_eta_live.py" = ["ANN002", "ANN003", "ANN401", "TC003", "E741"]
# dashboard_api carries Hermes-era webhook handlers + a long-lived `l`
# session-var idiom + req: Any framework hooks. Pre-existing debt to be
# cleaned up in a separate pass; not introduced by the canonical-path
# migration.
"deploy/scripts/dashboard_api.py" = ["ANN201", "ANN202", "ANN401", "E741", "E402", "E501"]
"deploy/scripts/cond_check.py" = ["E741", "E501", "TC003"]
"deploy/scripts/quick_check.py" = ["E741", "E402"]
"deploy/scripts/recent_verdicts.py" = ["E741", "TC003"]
# hermes_bridge embeds long Telegram webhook strings (E501) and FastAPI
# route handlers whose return types are framework-specified (ANN201/202).
"brain/jarvis_v3/hermes_bridge.py" = ["E501", "ANN201", "ANN202"]
"brain/jarvis_v3/quantum/qubo_supercharged.py" = ["N806"]
"feeds/_bump_roadmap_v0_1_14.py" = ["E501"]
"feeds/_bump_roadmap_v0_1_15.py" = ["E501"]
"feeds/_record_codebase_decision.py" = ["E501"]
"feeds/run_*_walk_forward.py" = ["ANN401"]
"scripts/run_*_walk_forward.py" = ["ANN401"]
"feeds/run_foundation_supercharge_sweep.py" = ["ANN401"]
"scripts/run_foundation_supercharge_sweep.py" = ["ANN401"]
"feeds/funding_ledger.py" = ["ANN401"]
"feeds/fleet_paper_launcher.py" = ["E501"]
"scripts/fleet_paper_launcher.py" = ["E501"]
"feeds/paper_trade_preflight.py" = ["E501"]
"scripts/paper_trade_preflight.py" = ["E501"]
"feeds/venue_readiness_check.py" = ["E501"]
"scripts/venue_readiness_check.py" = ["E501"]
"scripts/fleet_performance_report.py" = ["E501"]
"scripts/risk_dashboard.py" = ["E501"]
"scripts/sweep_reclaim_params.py" = ["E501"]
"scripts/tws_watchdog.py" = ["E501"]
"feeds/strategy_lab/app.py" = ["E501"]
"feeds/strategy_lab/run_fleet_sweep.py" = ["E402"]
"tests/live_force_multiplier.py" = ["E402"]
"scripts/jarvis_live.py" = ["E501"]
"deploy/scripts/ibkr_auth.py" = ["E402"]
"deploy/scripts/ibkr_bypass.py" = ["E402"]
"deploy/scripts/ibkr_full.py" = ["E402", "E501"]
"deploy/scripts/supercharge_fleet.py" = ["E501", "SIM102"]
"safety/execution_lease.py" = ["N818"]
# fetch_tws_historical_bars + its tests mirror the ib_insync API surface
# (Future / IB / qualifyContracts / reqHistoricalData with camelCase
# methods, camelCase params like clientId / endDateTime / barSizeSetting,
# and Any-typed contract/bar objects). N802/N803/ANN401 are the entire
# ib_insync ABI -- silencing them here is correct, not laziness.
"scripts/fetch_tws_historical_bars.py" = ["N802", "N803", "ANN401"]
"tests/test_fetch_tws_historical_bars.py" = ["N802", "N803"]
# Strategy-lab reports expose operator-facing expR/friction_R names.
"feeds/strategy_lab/engine.py" = ["N815"]
"feeds/strategy_lab/rigor.py" = ["N802", "N815"]
# Historical roadmap bump / decision ledger helpers intentionally embed long
# copied roadmap strings; wrapping them hurts auditability more than it helps.
"scripts/_bump_roadmap_v0_1_14.py" = ["E501"]
"scripts/_bump_roadmap_v0_1_15.py" = ["E501"]
"scripts/_record_codebase_decision.py" = ["E501"]
# per_bot_registry embeds long rationale strings in StrategyAssignment entries;
# wrapping them would hurt auditability.
"strategies/per_bot_registry.py" = ["E501"]
# paper_trade_sim has a long risk-sizing expression that can't be cleanly split.
"scripts/paper_trade_sim.py" = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
filterwarnings = [
    "ignore:'asyncio\\.get_event_loop_policy' is deprecated and slated for removal in Python 3\\.16:DeprecationWarning:eventkit\\.util",
]
markers = [
    "slow: deselect with '-m \"not slow\"' -- run-time-heavy tests (cross-regime sim, multi-bar walk-forward, etc.)",
]

[tool.mypy]
python_version = "3.12"
strict = true

